chatdoc/ingestion/requirements.txt
watchdog
openai
qdrant-client
PyMuPDF
python-docx
openpyxl
pytesseract
Pillow
